Create DocumentFinder to document methods
